+++ src/app/overlay/mod.rs
@@ -516,6 +516,7 @@ impl OverlayInner {
                 latest_combat,
                 combats: _,
                 difficulties: _,
+                base_names: _,
                 file_size: _,
             }) => latest_combat,
             _ => return,